backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pymysql==1.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
beautifulsoup4==4.12.3
requests==2.31.0
scrapy==2.11.0
pytest==8.0.0
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
